+++ doc/changelog.rst
@@ -6,6 +6,9 @@ ChangeLog
 0.2.0 (to be released)
 **********************
 
+* Rename ``gpgliblib.gnupg.GnuPGBackend`` to
+  :py:class:`gpgliblib.python_gnupg.PythonGnupgBackend` for consistency and to
+  avoid confusion with its fork, `gnupg <http://pythonhosted.org/gnupg/>`_.
 * Support for exporting keys via :py:func:`gpgliblib.base.GpgKey.export`.
 * Support for deleting keys via :py:func:`gpgliblib.base.GpgKey.delete`.
 * Add new attributes for GPG keys: